requirements/requirements-usage.txt
six>=1.10.0
graphene>=2.1.3,<3
graphql-core>=2.1.0,<3
Django>=1.11
djangorestframework>=3.9.0
singledispatch>=3.4.0.3
promise>=2.1